+++ src/main/resources/db/migration/V4__create_order_table.sql
@@ -1,7 +1,7 @@
 -- V4__create_order_table.sql
 CREATE TABLE Orders (
-                        id BIGINT PRIMARY KEY,
-                        user_id BIGINT NOT NULL,
+                        id SERIAL PRIMARY KEY,
+                        user_id INT NOT NULL,
                         order_state VARCHAR(20) NOT NULL,
                         total_amount DECIMAL(10, 2) NOT NULL,
                         editable BOOLEAN default true,
